+++ packages/mui-components/src/theme/defaultTheme.ts
@@ -132,4 +132,4 @@ export const createDefaultTheme = ({ darkTheme = false }: { darkTheme?: boolean 
   });
 };
 
-export const defaultTheme = createDefaultTheme({ darkTheme: true });
+export const defaultTheme = createDefaultTheme({ darkTheme: false });