+++ packages/oidc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @batoanng/oidc
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/mui-components@3.0.25
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/mui-components@3.0.24
+
 ## 0.0.20
 
 ### Patch Changes